+++ CHANGELOG.md
@@ -3,8 +3,12 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+* Add localization for sort order in list components.
+
 ### Removed
 * `ui-message` component: attribute `title`.
+
 ### Changed
 * `ui-message` component: `onShow` & `onHide` actions now sent on every change in visible property.