+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 * `flexberry-toogler` component:
     * Add `hasResizableOLV` flag indicates when component need to initialize `colResizable` plugin.
+* Add localization for sort order in list components.
 
 ### Removed
 * `ui-message` component: attribute `title`.