+++ CHANGELOG.md
@@ -3,17 +3,19 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.9.0-beta.0] - 2017-02-22
 ### Added
 * `flexberry-toogler` component:
     * Add `hasResizableOLV` flag indicates when component need to initialize `colResizable` plugin.
 * Add localization for sort order in list components.
 
-### Removed
-* `ui-message` component: attribute `title`.
-
 ### Changed
 * `ui-message` component: `onShow` & `onHide` actions now sent on every change in visible property.
+* Update dependency on `ember-flexberry-data` to version 0.9.0-beta.0.
 
+### Removed
+* `ui-message` component: attribute `title`.
 
 ## [0.8.6] - 2017-02-09
 ### Changed